requirements.txt
mkdocs-click
rich
lightning-grid
